libinput: less parallism to increase chances the test suite works
Discussion with the libinput maintainer has confirmed that many of the
libinput test cases need to run on a relatively unloaded system,
because input handling is by definition time-sensitive and any
event mismatches are considered a failure to avoid broken code hiding.

We can't expect libinput to run on an otherside unloaded machine
because the autobuilder ptests execute in a qemu at the same time
as builds can be happening.  We can however reduce the amount of
parallellism happening in the libinput test suite to give it a better
chance of succeeding.

This may not be enough, but it's worth a go.

[ YOCTO #14164 ]

(From OE-Core rev: c6f52b8e1cc98d9ea627017fab5103dbc2943ee0)

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>